fix tiny compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@6989 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/minidata/convhook.cpp
@@ -11,7 +11,7 @@ void InitMiniConvHook(int iJpegQuality)
 {
 	// specify conversion parameters
 	static convbase::MINI_CONVERSION_PARAMS conversion_params;
-	conversion_params.jpeg_quality = iJpegQuality; // jpeg quality in percent
+	conversion_params.jpeg_quality = (float) iJpegQuality; // jpeg quality in percent
 	conversion_params.usegreycstoration = false; // use greycstoration for image denoising
 	conversion_params.greyc_p=0.8f; // greycstoration sharpness, useful range=[0.7-0.9]
 	conversion_params.greyc_a=0.4f; // greycstoration anisotropy, useful range=[0.1-0.5]
